requirements.txt
python-telegram-bot[rate-limiter]>=21.0
selenium==4.23.1
webdriver-manager==4.0.2
python-dotenv==1.0.1
gunicorn==21.2.0
